ref: move createAddressObject to geohelpers
